# openapi.yaml
openapi: 3.0.0
info:
  title: OVS 管理平台 API
  version: 1.0.0
  description: |-
    Open vSwitch 管理平台后端 API，涵盖交换机、端口、Bond、流表、镜像、QoS、隧道、BFD、CFM、组播、STP/RSTP、NetFlow/sFlow/IPFIX、网络命名空间等全部常用与高级功能。
servers:
  - url: http://localhost:8080
paths:
...
  /api/ovs/bridge/set-datapath-type:
    post:
      summary: 设置 datapath_type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bridge:
                  type: string
                datapathType:
                  type: string
              required: [bridge, datapathType]
      responses:
        '200':
          description: 成功
...
  /api/ovs/port/add:
    post:
      summary: 新增端口
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bridge:
                  type: string
                portName:
                  type: string
                type:
                  type: string
              required: [bridge, portName]
      responses:
        '200':
          description: 成功
  /api/ovs/port/delete:
    post:
      summary: 删除端口
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bridge:
                  type: string
                portName:
                  type: string
              required: [bridge, portName]
      responses:
        '200':
          description: 成功
  /api/ovs/port/list:
    post:
      summary: 查询端口列表
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bridge:
                  type: string
              required: [bridge]
      responses:
        '200':
          description: 成功
...
  /api/ovs/port/set-vlan-mode:
    post:
      summary: 设置端口 VLAN mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                portName:
                  type: string
                vlanMode:
                  type: string
              required: [portName, vlanMode]
      responses:
        '200':
          description: 成功
  /api/ovs/port/set-trunks:
    post:
      summary: 设置端口 trunks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                portName:
                  type: string
                trunks:
                  type: array
                  items:
                    type: integer
              required: [portName, trunks]
      responses:
        '200':
          description: 成功
  /api/ovs/port/remove-property:
    post:
      summary: 移除端口属性
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                portName:
                  type: string
                property:
                  type: string
                value:
                  oneOf:
                    - type: integer
                    - type: string
                    - type: array
                      items:
                        type: integer
                    - type: array
                      items:
                        type: string
              required: [portName, property, value]
      responses:
        '200':
          description: 成功
  /api/ovs/port/info:
    post:
      summary: 查询端口属性
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                portName:
                  type: string
              required: [portName]
      responses:
        '200':
          description: 成功
  /api/ovs/patch/add:
    post:
      summary: 添加 Patch Port
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bridge:
                  type: string
                portName:
                  type: string
                peer:
                  type: string
              required: [bridge, portName, peer]
      responses:
        '200':
          description: 成功
  /api/ovs/tunnel/add:
    post:
      summary: 添加 Tunnel Port（GRE/Geneve/VXLAN）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bridge:
                  type: string
                portName:
                  type: string
                type:
                  type: string
                options:
                  type: object
                  additionalProperties:
                    type: string
              required: [bridge, portName, type]
      responses:
        '200':
          description: 成功
  /api/ovs/port/set-bfd:
    post:
      summary: 设置 BFD
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                portName:
                  type: string
                bfd:
                  type: object
                  additionalProperties:
                    type: string
              required: [portName, bfd]
      responses:
        '200':
          description: 成功
  /api/ovs/port/set-cfm:
    post:
      summary: 设置 CFM (802.1ag)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                portName:
                  type: string
                cfm:
                  type: object
                  additionalProperties:
                    type: string
              required: [portName, cfm]
      responses:
        '200':
          description: 成功
...
  /api/ovs/port/set-hfsc-qos:
    post:
      summary: 设置 HFSC QoS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                portName:
                  type: string
                maxRate:
                  type: string
                queues:
                  type: object
                  additionalProperties:
                    type: string
              required: [portName]
      responses:
        '200':
          description: 成功
  /api/ovs/port/updown:
    post:
      summary: 设置端口 up/down
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                netns:
                  type: string
                portName:
                  type: string
                up:
                  type: boolean
              required: [netns, portName, up]
      responses:
        '200':
          description: 成功
  /api/ovs/port/addr:
    post:
      summary: 给端口分配 IP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                netns:
                  type: string
                portName:
                  type: string
                ip:
                  type: string
              required: [netns, portName, ip]
      responses:
        '200':
          description: 成功
  /api/ovs/port/set-vlan:
    post:
      summary: 设置端口 VLAN tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                portName:
                  type: string
                tag:
                  type: integer
              required: [portName, tag]
      responses:
        '200':
          description: 成功
...